requirements.txt
aiohttp==3.8.1
aiohttp-apispec==2.2.3
aiosignal==1.2.0
apispec==3.3.2
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.0
frozenlist==1.3.1
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
multidict==6.0.2
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
webargs==5.5.3
yarl==1.8.1
